Updated to use legacy version
