--- lib/legacy/README
@@ -1,1 +1,0 @@
-This is the old version of Z3, used for legacy / backwards-compatibility purposes.
