+++ subprojects/solver/solver-z3-legacy/src/main/java/hu/bme/mit/theta/solver/z3legacy/Z3LegacySolverFactory.java
@@ -15,7 +15,7 @@
  */
 package hu.bme.mit.theta.solver.z3legacy;
 
-import com.microsoft.z3.InterpolationContext;
+import com.microsoft.z3legacy.InterpolationContext;
 import hu.bme.mit.theta.common.OsHelper;
 import hu.bme.mit.theta.solver.ItpSolver;
 import hu.bme.mit.theta.solver.Solver;
@@ -54,8 +54,8 @@ public final class Z3LegacySolverFactory implements SolverFactory {
 
     @Override
     public Solver createSolver() {
-        final com.microsoft.z3.Context z3Context = new com.microsoft.z3.Context();
-        final com.microsoft.z3.Solver z3Solver = z3Context.mkSimpleSolver();
+        final com.microsoft.z3legacy.Context z3Context = new com.microsoft.z3legacy.Context();
+        final com.microsoft.z3legacy.Solver z3Solver = z3Context.mkSimpleSolver();
 
         final Z3SymbolTable symbolTable = new Z3SymbolTable();
         final Z3TransformationManager transformationManager = new Z3TransformationManager(
@@ -68,8 +68,8 @@ public final class Z3LegacySolverFactory implements SolverFactory {
 
     @Override
     public UCSolver createUCSolver() {
-        final com.microsoft.z3.Context z3Context = new com.microsoft.z3.Context();
-        final com.microsoft.z3.Solver z3Solver = z3Context.mkSimpleSolver();
+        final com.microsoft.z3legacy.Context z3Context = new com.microsoft.z3legacy.Context();
+        final com.microsoft.z3legacy.Solver z3Solver = z3Context.mkSimpleSolver();
 
         final Z3SymbolTable symbolTable = new Z3SymbolTable();
         final Z3TransformationManager transformationManager = new Z3TransformationManager(
@@ -83,7 +83,7 @@ public final class Z3LegacySolverFactory implements SolverFactory {
     @Override
     public ItpSolver createItpSolver() {
         final InterpolationContext z3Context = InterpolationContext.mkContext();
-        final com.microsoft.z3.Solver z3Solver = z3Context.mkSimpleSolver();
+        final com.microsoft.z3legacy.Solver z3Solver = z3Context.mkSimpleSolver();
 
         final Z3SymbolTable symbolTable = new Z3SymbolTable();
         final Z3TransformationManager transformationManager = new Z3TransformationManager(
