+++ subprojects/solver/solver-z3-legacy/src/main/java/hu/bme/mit/theta/solver/z3legacy/Z3SymbolTable.java
@@ -26,7 +26,7 @@ import static com.google.common.base.Preconditions.checkState;
 
 final class Z3SymbolTable {
 
-    private final BiMap<ConstDecl<?>, com.microsoft.z3.FuncDecl> constToSymbol;
+    private final BiMap<ConstDecl<?>, com.microsoft.z3legacy.FuncDecl> constToSymbol;
 
     public Z3SymbolTable() {
         constToSymbol = Maps.synchronizedBiMap(HashBiMap.create());
@@ -36,22 +36,22 @@ final class Z3SymbolTable {
         return constToSymbol.containsKey(constDecl);
     }
 
-    public boolean definesSymbol(final com.microsoft.z3.FuncDecl symbol) {
+    public boolean definesSymbol(final com.microsoft.z3legacy.FuncDecl symbol) {
         return constToSymbol.inverse().containsKey(symbol);
     }
 
-    public com.microsoft.z3.FuncDecl getSymbol(final ConstDecl<?> constDecl) {
+    public com.microsoft.z3legacy.FuncDecl getSymbol(final ConstDecl<?> constDecl) {
         checkArgument(definesConst(constDecl), "Declaration %s not found in symbol table",
                 constDecl);
         return constToSymbol.get(constDecl);
     }
 
-    public ConstDecl<?> getConst(final com.microsoft.z3.FuncDecl symbol) {
+    public ConstDecl<?> getConst(final com.microsoft.z3legacy.FuncDecl symbol) {
         checkArgument(definesSymbol(symbol), "Symbol %s not found in symbol table", symbol);
         return constToSymbol.inverse().get(symbol);
     }
 
-    public void put(final ConstDecl<?> constDecl, final com.microsoft.z3.FuncDecl symbol) {
+    public void put(final ConstDecl<?> constDecl, final com.microsoft.z3legacy.FuncDecl symbol) {
         checkNotNull(constDecl);
         checkNotNull(symbol);
         checkState(!constToSymbol.containsKey(constDecl), "Constant %s not found.", constDecl);
