+++ subprojects/solver/solver-z3-legacy/src/main/java/hu/bme/mit/theta/solver/z3legacy/Z3TypeTransformer.java
@@ -16,7 +16,7 @@
 package hu.bme.mit.theta.solver.z3legacy;
 
 import com.google.common.collect.Sets;
-import com.microsoft.z3.Context;
+import com.microsoft.z3legacy.Context;
 import hu.bme.mit.theta.core.type.Type;
 import hu.bme.mit.theta.core.type.arraytype.ArrayType;
 import hu.bme.mit.theta.core.type.booltype.BoolType;
@@ -35,11 +35,11 @@ final class Z3TypeTransformer {
     private final Z3TransformationManager transformer;
     private final Context context;
 
-    private final com.microsoft.z3.BoolSort boolSort;
-    private final com.microsoft.z3.IntSort intSort;
-    private final com.microsoft.z3.RealSort realSort;
-    private final Set<com.microsoft.z3.BitVecSort> bvSorts;
-    private final Set<com.microsoft.z3.FPSort> fpSorts;
+    private final com.microsoft.z3legacy.BoolSort boolSort;
+    private final com.microsoft.z3legacy.IntSort intSort;
+    private final com.microsoft.z3legacy.RealSort realSort;
+    private final Set<com.microsoft.z3legacy.BitVecSort> bvSorts;
+    private final Set<com.microsoft.z3legacy.FPSort> fpSorts;
 
     Z3TypeTransformer(final Z3TransformationManager transformer, final Context context) {
         this.context = context;
@@ -52,7 +52,7 @@ final class Z3TypeTransformer {
         fpSorts = Sets.synchronizedNavigableSet(new TreeSet<>());
     }
 
-    public com.microsoft.z3.Sort toSort(final Type type) {
+    public com.microsoft.z3legacy.Sort toSort(final Type type) {
         if (type instanceof BoolType) {
             return boolSort;
         } else if (type instanceof IntType) {
@@ -61,33 +61,33 @@ final class Z3TypeTransformer {
             return realSort;
         } else if (type instanceof BvType) {
             final BvType bvType = (BvType) type;
-            final Optional<com.microsoft.z3.BitVecSort> bvSort = bvSorts.stream()
+            final Optional<com.microsoft.z3legacy.BitVecSort> bvSort = bvSorts.stream()
                     .filter(sort -> sort.getSize() == bvType.getSize()).findAny();
             if (bvSort.isPresent()) {
                 return bvSort.get();
             } else {
-                final com.microsoft.z3.BitVecSort newBvSort = context.mkBitVecSort(
+                final com.microsoft.z3legacy.BitVecSort newBvSort = context.mkBitVecSort(
                         bvType.getSize());
                 bvSorts.add(newBvSort);
                 return newBvSort;
             }
         } else if (type instanceof FpType) {
             final FpType fpType = (FpType) type;
-            final Optional<com.microsoft.z3.FPSort> fpSort = fpSorts.stream().filter(
+            final Optional<com.microsoft.z3legacy.FPSort> fpSort = fpSorts.stream().filter(
                     sort -> sort.getEBits() == fpType.getExponent()
                             && sort.getSBits() == fpType.getSignificand()).findAny();
             if (fpSort.isPresent()) {
                 return fpSort.get();
             } else {
-                final com.microsoft.z3.FPSort newFpSort = context.mkFPSort(fpType.getExponent(),
+                final com.microsoft.z3legacy.FPSort newFpSort = context.mkFPSort(fpType.getExponent(),
                         fpType.getSignificand());
                 fpSorts.add(newFpSort);
                 return newFpSort;
             }
         } else if (type instanceof ArrayType) {
             final ArrayType<?, ?> arrayType = (ArrayType<?, ?>) type;
-            final com.microsoft.z3.Sort indexSort = toSort(arrayType.getIndexType());
-            final com.microsoft.z3.Sort elemSort = toSort(arrayType.getElemType());
+            final com.microsoft.z3legacy.Sort indexSort = toSort(arrayType.getIndexType());
+            final com.microsoft.z3legacy.Sort elemSort = toSort(arrayType.getElemType());
             return context.mkArraySort(indexSort, elemSort);
         } else {
             throw new UnsupportedOperationException(
